Move FFE helpers to canonical utils module
